agents/pyproject.toml
[tool.poetry]
name = "agents"
version = "0.1.0"
description = "Small, efficient AI agents for automation"
authors = ["Your Name <landonmutch@protonmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
python-dotenv = "^1.0.0"
pytz = "^2024.1"
smolagents = "^1.10.0"
markdownify = "^0.14.1"
requests = "^2.32.3"
litellm = "^1.61.15"
opentelemetry-exporter-otlp = "^1.30.0"
openinference-instrumentation-smolagents = "^0.1.6"
opentelemetry-sdk = "^1.30.0"
arize-phoenix = "^8.2.1"
google-genai = "^1.3.0"
google-generativeai = "^0.8.4"
arxiv = "^2.1.0"
pymupdf4llm = "^0.0.17"
uuid = "^1.30"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 